+++ test/model_tests/debate_event_test.dart
@@ -9,17 +9,39 @@ void main() {
     setUp(() {
       debateEvent = Policy.highSchool();
     });
-    test('The speeches getter returns a non-null list of speeches.', () {
+    test('get speeches returns a non-null list of speeches.', () {
       expect(debateEvent.speeches, isNotNull);
       expect(debateEvent.speeches is List<Speech>, isTrue);
     });
-    test('Calling nextSpeech() increments the speech index.', () {});
-    test('Calling nextSpeech() on last index throws IndexError.', () {});
-    test('Calling prevSpeech() decrements the speech index.', () {});
-    test('Calling prevSpeech() on first index throws IndexError.', () {});
+    test('nextSpeech() increments the speech index.', () {
+      expect(debateEvent.speeches.indexOf(debateEvent.speech), equals(0));
+      debateEvent.nextSpeech();
+      expect(debateEvent.speeches.indexOf(debateEvent.speech), equals(1));
+    });
+    test('nextSpeech() throws IndexError on last index.', () {
+      for (int i = 0; i < debateEvent.speeches.length - 1; i++) {
+        debateEvent.nextSpeech(); // skip through all but the last speech
+      }
+      expect(() => debateEvent.nextSpeech(), throwsRangeError);
+    });
+    test('prevSpeech() decrements the speech index.', () {
+      debateEvent.nextSpeech(); // to avoid throwing RangeError
+      expect(debateEvent.speeches.indexOf(debateEvent.speech), equals(1));
+      debateEvent.prevSpeech();
+      expect(debateEvent.speeches.indexOf(debateEvent.speech), equals(0));
+    });
+    test('prevSpeech() throws IndexError on first index.', () {
+      expect(() => debateEvent.prevSpeech(), throwsRangeError);
+    });
     test('dispose() clears the list of speeches.', () {
       debateEvent.dispose();
       expect(debateEvent.speeches, isEmpty);
+      debateEvent = Policy.highSchool(); // set back up for teardown
+    });
+    test('dispose() clears the Event speech.', () {
+      debateEvent.dispose();
+      expect(debateEvent.speech, isNull);
+      debateEvent = Policy.highSchool(); // set back up for teardown
     });
     tearDown(() {
       debateEvent.dispose();