+++ test/model_tests/prep_time_mixin_test.dart
@@ -1,0 +1,42 @@
+import 'package:preptime/provider/models/prep_time_mixin.dart';
+import 'package:preptime/provider/models/speech_event.dart';
+import 'package:test/test.dart';
+
+/// This is probably an anti-pattern.
+///
+/// The [PrepTimeMixin] should really only be defined for [DebateEvents], but
+/// technically it can be mixed in on any [Event]. Therefore, to prevent
+/// isolate the [PrepTimeMixin] for testing, we will create this
+/// [TestPrepTimeMixin] class and then go ahead and not use any of the
+/// [SpeechEvent] parts of the parent.
+class TestPrepTimeMixin extends SpeechEvent with PrepTimeMixin {}
+
+void main() {
+  group('PrepTimeMixin', () {
+    TestPrepTimeMixin prepTimeMixin;
+    setUp(() {
+      prepTimeMixin = TestPrepTimeMixin();
+    });
+    test('initialPrep is initially null', () {
+      expect(prepTimeMixin.initialPrep, isNull);
+    });
+    test('initialPrep equals duration after initPrepTimers is called', () {
+      Duration _duration = const Duration(minutes: 3);
+      expect(prepTimeMixin.initialPrep, isNull);
+      prepTimeMixin.initPrepTimers(duration: _duration);
+      expect(prepTimeMixin.initialPrep, equals(_duration));
+    });
+    test('initTimers constructs ', () {});
+    test('disposePrepTimers', () {});
+    test('isRunning', () {});
+    test('isNotRunning', () {});
+    test('isOtherRunning', () {});
+    test('startPrep', () {});
+    test('stopPrep', () {});
+    test('togglePrep', () {});
+    test('resetPrep', () {});
+    test('remainingPrep', () {});
+    test('prepName', () {});
+    tearDown(() {});
+  });
+}